Remove support for passing a C-String to PKCS5_PBKDF2_HMAC*().
Before it was possible to pass a NULL-terminated C-string to the PBKDF2
functions, and indicate the parameter was a C-string by passing a length
of -1.

This is not relied on anywhere in the BoringSSL code, and the API contract is
possible to misuse as it is not the common way of doing this.

(Passing in an large unsigned length and having it be interpreted as -1,
and consequently a C-string)

Change-Id: Ifbd31ff76e183fa74e9fa346908daf4bfb8fc3da