Initialize HMAC keys to zero.
It is easy for consumers to inadvertently call HMAC_Init_ex(key=NULL, key_len=0) in an attempt to assign a zero-length HMAC key. This does not work as expected because key==NULL has special semantics, and can result in the use uninitialized memory as the key data.

This workaround default initializes HMAC keys to zero-length, so this misuse of the API no longer results in the use of uninitialized memory.

BUG=http://crbug.com/449409

Change-Id: I777276d57c61f1c0cce80b18e28a9b063784733f