Update patch set 4
Patch Set 4:

(6 comments)

The problem with this is that you would be the only person building on Windows with the CMake build I suspect, and thus we'll probably be constantly breaking it.

The fact that Windows seemingly can't import its own headers with this level of warnings suggests that it might be too much. Each line of preprocessor has quite a high cost.

Patch-set: 4
Label: Code-Review=0