Update patch set 4
Patch Set 4:

(1 comment)

I adjusted my patch to address your review comments.

Regarding your concern that this will cause the Windows CMake build to often be broken: It's not officially supported anyway, and I'm happy to fix the breakage. Also, if I'm indeed the only user, it doesn't matter much. If/when breakage happens, we can also enable the analogous GCC & Clang warnings to reduce the changes of platform-specific breakage in the future.

Without these changes, it's more likely that somebody developing on Windows will make a change that breaks the Clang, GCC, and/or any of the Chromium builds, since those platforms already enable many of the same warnings.

Patch-set: 4
Label: Code-Review=0