Update patch set 4
Patch Set 4:

Eh. As I noted in the original PR, what's going on in a_object.c is very unclear to me. I suspect that it's most likely that for some inputs, the function is wrong. But it's so complex, it's really quite hard to see how it is correct or incorrect.

Patch-set: 4
Reviewer: Gerrit User 5315 <5315@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0