Update patch set 7
Patch Set 7:

> Do you have a Chromium-side change that goes along with this?

I've uploaded an experimental patch here. https://codereview.chromium.org/916183002 
This is work in progress and It's filled with global variables and other memory accounting code that is not supposed to be in a final patch. 

The main idea is to remove the buffers if the socket goes idle after the handshake is finished. That does not seem to considerable raise the complexity and it has a good return as chrome opens a lot of idle sockets.
 
In this patch I've gone a bit further and also removed the buffers if the socket goes idle after usage. I also remove the write buffers if the socket hangs on read (which is a common technique used by servers to push events to the browser). It seems to work without trouble but there must be some guaranty that the buffers isn't needed for other reasons than read and write.

In the log "all socket memory saved:" reports total memory savings due to both zero copy api and buffer removal. "New socket memory saved" reports buffer removal only. I typically see between 500Kb to 1Mb savings on loading a https site, but it depends a lot of the complexity of the sites.

Patch-set: 7