Update patch set 2
Patch Set 2: Code-Review+2

Dear god. There's no way that I can convince myself that the change in tasn_new.c is safe. asn1_item_ex_combine_new is corecursive with some of the functions that it calls! I'm just going by the fact that auxerr does it for that one.

Patch-set: 2
Label: Code-Review=+2