Update patch set 1
Patch Set 1:

(Also confirmed against a test server that this CL is sufficient. I'll go patch runner.go to do the full ephemeral key exchange so our tests are slightly more complete. We reject the ServerKeyExchange early and the error code should be sufficient, but just to be thorough.)

Patch-set: 1
Label: Code-Review=0