Update patch set 9
Patch Set 9:

> (12 comments)
 > 
 > Gah, sorry, I really need to get better about review turnaround
 > time.
 > 
 > Re Chromium: To confirm, this is really only useful for idle
 > HTTP/1.1 sockets in the socket pool, right? I believe SPDY (or
 > HTTP/2) sessions have a hanging Read steady-state.

It is possible to release the write buffers only on hanging reads. If you have a close look at the chromium patch, I've done exactly that, although I'm not sure that will end up in the final patch. It depends a little on how much is saved through those, if it is worth the complexity, and if you allow it :)

Patch-set: 9