Generate error data at build time.
This avoids cluttering up the diff and making merge conflicts a pain.  It does,
however, mean we need to generate err_data.c ahead of time in Chromium and
likely other downstream builds. It also adds a build dependency on Go.

[Not to land quite yet; I want to make sure I have the update_gypi_and_asm.py
and recipe changes ready to go first so we don't block rolls.]

Change-Id: I6e0513ed9f50cfb030f7a523ea28519590977104