Fix incorrect assert.
This assert came from a test in the OpenSSL code that recorded an
internal error. Thanks to Sean Burford for finding that it's possible to
trigger this assert.

Thankfully, it's the assert that appears to be incorrect rather than the
rest of the code as it accounts for any existing data in the buffer
twice.

In practice, this assert will be compiled out anyway.

Change-Id: Iad070cf9ed1d9be7857e12b5701f36e3a43195e0