Update patch set 3
Patch Set 3: Code-Review+2

Hmm, those underlying DES functions look like they'll crash if the input length isn't a multiple of 8 now that len is a size_t. Since they return void, that might actually be the safest option I guess but then the code for handling inputs that are not a multiple of 8 should be removed. Will poke.

Patch-set: 3
Label: Code-Review=+2