Update patch set 1
Patch Set 1:

> Why? The doc generation still seems to work.

Sorry. When I wrote this patch, I was using Windows, and there's no indication on Windows that this is a symlink.

I will abandon this changeset. I think it would be useful to change the doc generation so that this symlink isn't necessary. In particular, sometime it should become possible to generate the docs into a directory other than the source directory.

Patch-set: 1
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0