Update patch set 2
Abandoned

This is the end-goal of the other changes I submitted recently, but I submitted this one too early. There are more changes needed to remove all the dependencies on asn1 & obj from the low-level crypto stuff that I have in my local tree, but that I'd forgotten about.

Patch-set: 2
Status: abandoned