Update patch set 2
Patch Set 2:

Then, AFAICT, we would be screwed if we ever made a mistake and executed that code path expecting it to work. The abort() calls act as a stopgap to prevent such mistakes from being catastrophic.

There are a lot of warnings (especially the MSVC /analyze ones, and clang static analysis warnings) that would be very useful for ensuring that BoringSSL is (memory-)safe, but which require some annotations to handle false positives. I think we could generalize the pattern established in patch set 2 to reasonably deal with that scenerio. The alternative seems to be to just forgo the benefit of those static analysis tools, which I think would be unfortunate.

Patch-set: 2
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0