Update patch set 3
Patch Set 3:

> > Just some print hooks to extract
 > > which would probably be tidier.
 > 
 > Please let me know if the changes I just submitted are what you are
 > referring to here, or whether you are referring to something else.

Oh, sorry, that was unclear. I meant that, after the core crypto bits don't depend on crypto/asn1, it looks like the only additional dependency left on BIO is the print hooks in p_rsa_asn1.c and friends. If we wanted to go through with removing the BIO dependency, we'd just have extracting (or removing? haven't checked on usage) those left, and that would probably be tidier [than what we have now] anyway. So not a comment that your CL left something out but a comment on some other stuff that would also need to be extracted.

Patch-set: 3
Reviewer: Gerrit User 5075 <5075@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0