Explicitly document EVP_PKEY_size vs EVP_PKEY_bits.
Notably, they do not return the same value for EC keys. "size" is a really
weird notion to generalize on so we should document what it means for each key
type. EVP_PKEY_size's meaning is most tied to signatures, thanks to
EVP_SignFinal implicitly using it as output bounds.

Change-Id: I7504c142818f8f90f8bcf6891c97a6adaf2d574e