Update patch set 3
Patch Set 3:

Turns out most of find_depot_tools.py was kind of pointless. I've trimmed it down to the minimum that we need.

This could probably be trimmed down considerably more (the recipe knows the actual path to depot_tools and such), but I've left it as-is for now. There's some chance we'll need to re-sync from Chromium later, so there may be something to be said for not diverging. Apparently we don't really have a Windows toolchain story outside of Chromium's mess so everything else is dubiously supported.

Patch-set: 3
Reviewer: Gerrit User 5075 <5075@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0