Update patch set 2
Patch Set 2:

> If that's the concern, we'd want to preserve state across
 > non-blocking calls, right? I could just wait a bit between each
 > burst of warning alerts. (Though I suppose that throttles it.)

Keeping state across non-blocking unwinds is better as long as the complexity is small enough. Not doing it still helps because an attacker would have to be careful not to trigger it.

The empty-records check was done to be as simple as possible so that upstream might take it without ABI worries. Now, we can likely do better.

Patch-set: 2