Update patch set 1
Patch Set 1:

(1 comment)

If this is just to remove the dependency on ASN.1 parsing, the bssl tool is always going to need some form of ASN.1 parsing by virtue of the SSL bits. And unless Google wants it for something, we're not likely to want to support an OPENSSL_NO_ASN1 build configuration; build knobs are a rather annoying burden.

When I said I'd like to rid ourselves of crypto/asn1, I meant implement ASN.1 parsing with crypto/bytestring, not remove the dependencies on ASN.1 altogether. And some amount of that is premature right now because we don't know what our certificate verifier is going to look like. (E.g. I'm probably going to want to teach ssl/ to treat certificates as opaque byte strings and maybe just know the bare minimum needed to extract a public key. I'm not sure. But it's really not on my radar until the certificate verifier is ready and we know, e.g., whose job it is parsing RSAPublicKeys, SubjectPublicKeyInfo, etc., and whether it's more convenient to have callback or what.)

Patch-set: 1