Update patch set 1
Patch Set 1:

> If this is just to remove the dependency on ASN.1 parsing, the bssl
 > tool is always going to need some form of ASN.1 parsing by virtue
 > of the SSL bits. And unless Google wants it for something, we're
 > not likely to want to support an OPENSSL_NO_ASN1 build
 > configuration; build knobs are a rather annoying burden.

Understood. I wrote this patch for my own use (along with literally 100 others) and this is one that I thought you may find useful. In my case, removing the dependency on d2i_RSAPrivateKey is useful. I recognize that removing that dependency may not be so interesting to Google. I am not sure if the improved readability of the speed.cc code makes it worthwhile to you or not. It is fine by me if you want to abandon this changeset.

 > (E.g. I'm probably going to want to
 > teach ssl/ to treat certificates as opaque byte strings and maybe
 > just know the bare minimum needed to extract a public key. I'm not
 > sure. But it's really not on my radar until the certificate
 > verifier is ready and we know, e.g., whose job it is parsing
 > RSAPublicKeys, SubjectPublicKeyInfo, etc., and whether it's more
 > convenient to have callback or what.)

I've already done a draft set of modifications to libssl to remove its dependencies on crypto/asn1, crypto/x509, and crypto/x509v3 in my tree, though they're not ready to share yet. Your ideas mention seem to be on the right track.

Patch-set: 1
Label: Code-Review=0