Update patch set 3
Patch Set 3:

With a variable-time implementation of either the sign and/or verify function, the results would vary based on the bits of the keys used. Thus, it makes more sense to always use the same keypair for each run so that results are comparable. It's still possible to do that while avoiding the d2i_* functions, but the current patch doesn't do that. I will write a new one soon.

Patch-set: 3
Reviewer: Gerrit User 5525 <5525@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0