Update patch set 2
Patch Set 2:

> This area of the code is so full of terrors that I suspect that
 > something will go wrong, but I can't see it at the moment.

Well, there's going to be more violence done to it later. :-) I have a branch that replaces tls1_enc with SSL_AEAD_CTX_open and SSL_AEAD_CTX_seal that take the usual out/out_len/max_out trio and removes s->s3->wrec (though I'm thinking it might be worth swapping that for CBB; the overflow and bounds checks are tedious).

It may also be worth seeing about replacing the memcpy-for-alignment-and-historical-compression-remnants + in-place seal with a non-in-place seal. I'll have to do some bssl speed runs to see whether there's any difference (and if it is faster, the memcpy probably ought to be moved behind EVP_AEAD_seal). In the non-record-splitting case, we may also just be able to lean on the caller supplying aligned input when it matters.

Besides, before this CL, the record-splitting case didn't get the alignment right anyway.

Patch-set: 2
Reviewer: Gerrit User 5075 <5075@66b711a4-aa9b-3dac-9062-916540830212>
Label: Code-Review=0